Moved Grade and SimpleGrade to model package. Tested SimpleGrade.
